modifiche PuntoService.java

+++ idsProject/src/main/java/com/cs/idsProject/service/PuntoService.java
@@ -10,8 +10,10 @@ import java.util.Optional;
 
 @Service
 public class PuntoService {
+
     @Autowired
     private PuntoRepository puntoRepository;
+
 
     public List<POI> getAllPunti() {
         return puntoRepository.findAll();
@@ -21,11 +23,27 @@ public class PuntoService {
         return puntoRepository.findById(id);
     }
 
-    public POI addPunto(POI punto) {
+    public POI addPunto(Integer idComune, POI punto) {
+        // Verifica se il punto esiste già per lo stesso comune
+        Optional<POI> existingPunto = puntoRepository.findByNomeAndLatitudineAndLongitudineAndIdComune(
+                punto.getNome(), punto.getLatitudine(), punto.getLongitudine(), idComune);
+
+        if (existingPunto.isPresent()) {
+            throw new IllegalArgumentException("Il punto con lo stesso nome e coordinate esiste già nel comune specificato.");
+        }
+
+        punto.setIdComune(idComune);
         return puntoRepository.save(punto);
     }
 
-    public void deletePunto(int id) {
-        puntoRepository.deleteById(id);
+    public void deletePunto(Integer id, Integer idComune) {
+        // Verifica se il punto esiste nel repository per lo stesso comune
+        Optional<POI> existingPunto = puntoRepository.findByIdAndIdComune(id, idComune);
+
+        if (existingPunto.isPresent()) {
+            puntoRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("Il punto con l'ID e il comune forniti non esiste.");
+        }
     }
 }
